add num of noti
